feat: remove uuid crate dependency
`uuid` pulls in a bunch of WASM dependencies, even when Codspeed is pulled in with `default-features = false`. For some projects, that means vetting a lot of extra code.

This PR replaces the `uuid` dependency and uses `getrandom` to generate an UUIDv4-style string. (I'm not sure if anything depends on things being named with UUIDs; if not, this PR would be even simpler.)